Add files via upload
Question 1: Create a basic HTML5 page with a title, heading, paragraph, and an image.
In this file, I have created a simple HTML5 page about my favourite hobby which is reading. The page includes a title, a main heading, a descriptive paragraph about my hobby along with an image of a stack of books.